调整跑胡子目录
