+++ phz/phzlib_c/hulib.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int get_huxi(char* cards);
