+++ phz/phzlib_lua/gen_eye_table.bat
@@ -1,0 +1,2 @@
+lua.exe gen_eye_table.lua
+pause
