+++ phz/phzlib_lua_jiao/gen_table.bat
@@ -1,0 +1,2 @@
+lua.exe gen_table.lua
+pause
