--- phzlib_c/hulib.h
@@ -1,3 +1,0 @@
-#pragma once
-
-int get_huxi(char* cards);
