--- phzlib_lua/gen_eye_table.bat
@@ -1,2 +1,0 @@
-lua.exe gen_eye_table.lua
-pause
