--- phzlib_lua/gen_table.bat
@@ -1,2 +1,0 @@
-lua.exe gen_table.lua
-pause
